add dashboard tabs, fix front styling, change home image
