+++ app/assets/stylesheets/components/_album_card.scss
@@ -142,9 +142,8 @@
 
 .album-card-minimal:hover .far {
   display: block;
+  text-shadow: 1px 1px 3px rgba(0,0,0,0.6);
 }
-
-
