+++ app/assets/stylesheets/pages/_dashboard.scss
@@ -39,4 +39,5 @@
 .tabs-selector:hover {
   padding: 10px;
   text-decoration: none;
+  cursor: pointer;
 }
