+++ app/javascript/packs/application.js
@@ -28,35 +28,13 @@ import "bootstrap";
 // Internal imports, e.g:
 // import { initSelect2 } from '../components/init_select2';
 import { initAutocomplete } from '../plugins/init_autocomplete';
-
-
 import { initMapbox } from '../plugins/init_mapbox';
+import { tabsSelection } from '../tabs';
 
 document.addEventListener('turbolinks:load', () => {
   // Call your functions here, e.g:
   // initSelect2();
   initAutocomplete();
   initMapbox();
-});
-
-const tabs = document.querySelectorAll(".tabs-selector");
-
-tabs.forEach((tab) => {
-  tab.addEventListener("click", (event) => {
-    event.preventDefault();
-    const currentActiveTabSelector = document.querySelector(".tabs-selector-active");
-    const currentInactiveTabSelector = document.querySelector(".tabs-selector-inactive");
-    const currentActiveTab = document.querySelector(".active-tab");
-    const currentInactiveTab = document.querySelector(".inactive-tab");
-    if (currentActiveTabSelector !== event.currentTarget) {
-      currentActiveTabSelector.classList.toggle("tabs-selector-active");
-      currentActiveTabSelector.classList.toggle("tabs-selector-inactive");
-      currentInactiveTabSelector.classList.toggle("tabs-selector-active");
-      currentInactiveTabSelector.classList.toggle("tabs-selector-inactive");
-      currentActiveTab.classList.toggle("active-tab");
-      currentActiveTab.classList.toggle("inactive-tab");
-      currentInactiveTab.classList.toggle("active-tab");
-      currentInactiveTab.classList.toggle("inactive-tab");
-    };
-  });
+  tabsSelection();
 });
